Unify test implementations
