+++ interop-tests/Cargo.toml
@@ -60,6 +60,5 @@ wasm-bindgen = { version = "0.2" }
 wasm-bindgen-futures = { version = "0.4" }
 wasm-logger = { version = "0.2.0" }
 wasm-timer = { version = "0.2.5" }
-gloo-timers = { version = "0.2.6" }
 reqwest = { version = "0.11", features = ["json"] }
 console_error_panic_hook = { version = "0.1.7" }
